refactor(capability): keep the two megafiles under their gate while the fix lands
check:filesize caught both files this change touches. Neither is trimmed by
deleting the fix:

- integrationSession.ts: both credential-ready writes shared a five-line commit
  tail, now one `commitCredentialReady` that also carries the handoff retirement,
  so the retirement has exactly one home instead of two copies.
- capabilityApplyHandler.ts: #528's two renderer-only host effects move to
  `mcpHostSurfaceOps`, the same "落点住在别处，本 handler 保持精简" split the file
  already uses for multiShotCanvasLanding. They pass 800 with room, not on it.

OnboardingDrawer and integrationSession baselines drop to what they now measure.
