+++ electron/integrationCertification/integrationSession.ts
@@ -140,7 +140,6 @@ type Dependencies = {
     >;
   /** Durable UI handoff sink. The session service never emits an event-only handoff. */
   enqueueHandoff?: (input: Omit<IntegrationHandoff, "requestId" | "createdAt">) => unknown;
-  /** Retires a durable handoff whose reason to exist is gone, whichever route resolved it. */
   retireHandoff?: (sessionId: string, target: IntegrationHandoff["target"]) => unknown;
   now?: () => string;
   /** Durable reservation for native ComfyUI certification submissions. */
@@ -153,7 +152,6 @@ export function createRuntimeIntegrationSessionService(
     approvalReceiptAuthority?: Dependencies["approvalReceiptAuthority"];
     certification?: ConnectionCertificationService;
     enqueueHandoff?: Dependencies["enqueueHandoff"];
-    retireHandoff?: Dependencies["retireHandoff"];
     save?: Dependencies["save"];
     filePath?: string;
     now?: () => string;
@@ -470,7 +468,7 @@ export function createRuntimeIntegrationSessionService(
     certification,
     approvalReceiptAuthority: authority,
     enqueueHandoff: input.enqueueHandoff || enqueueIntegrationHandoff,
-    retireHandoff: input.retireHandoff || retireIntegrationHandoffs,
+    retireHandoff: retireIntegrationHandoffs,
     save: input.save,
     now: input.now,
     credentialResolver: resolveCredential,
@@ -1104,22 +1102,20 @@ export class IntegrationSessionService {
     });
     return result;
   }
-  markCredentialReady(
-    sessionId: string,
-    credentialRef: string,
-    owner: CapabilityOriginHost,
-  ): IntegrationSessionProjection {
+  markCredentialReady(sessionId: string, credentialRef: string, owner: CapabilityOriginHost): IntegrationSessionProjection {
     const session = this.getOrThrow(sessionId);
     if (session.ownerClientId !== owner || owner === "external") throw new Error("Signed client identity is required");
     session.credentialRef = id(credentialRef, "credentialRef");
     session.credentialStatus = "ready";
     session.stage = "draft";
+    return this.commitCredentialReady(session);
+  }
+  /** Shared tail of both credential-ready writes. The "type a key" handoff is retired only after the write is on disk. */
+  private commitCredentialReady(session: IntegrationSession): IntegrationSessionProjection {
     session.revision += 1;
     session.updatedAt = (this.deps.now || (() => new Date().toISOString()))();
     this.state.revision += 1;
     this.persist();
-    // Only after the credential is durably recorded. Retiring first would drop the "type a key"
-    // request on a write that then failed, leaving the user with no route back to the page.
     this.deps.retireHandoff?.(session.id, "credential");
     return this.projection(session);
   }
@@ -1167,12 +1163,7 @@ export class IntegrationSessionService {
     session.credentialStatus = "ready";
     session.stage = "draft";
     session.blockingReason = undefined;
-    session.revision += 1;
-    session.updatedAt = (this.deps.now || (() => new Date().toISOString()))();
-    this.state.revision += 1;
-    this.persist();
-    this.deps.retireHandoff?.(session.id, "credential");
-    return this.projection(session);
+    return this.commitCredentialReady(session);
   }
   async propose(
     sessionId: unknown,
